fix(dashboard,crm): revenue units, lead normalization, date-ranged us…

+++ Core/Services/Dashboard/Contracts/GetOverallSummaryDto.cs
@@ -6,6 +6,7 @@ public record GetOverallSummaryDto
     public double TotalUsersGrows { get; set; }
     public int TotalSalesCount { get; set; }
     public double TotalSalesCountGrows { get; set; }
-    public long TotalSalesAmount { get; set; }
+    // So'mda (Amount DB'da tiyinda saqlanadi — /100 qilib qaytariladi).
+    public double TotalSalesAmount { get; set; }
     public double TotalSalesAmountGrows { get; set; }
 }
